+++ src/shared/ui/Toaster/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Toaster';
